src/api-gateway/requirements.txt
fastapi==0.115.12
httpx==0.28.1
PyYAML==6.0.2
uvicorn==0.34.0